refactor: lab info and service
