+++ containers/ecr-viewer/src/app/services/ecrSummaryService.tsx
@@ -21,13 +21,12 @@ import {
   returnImmunizations,
   returnProblemsTable,
 } from "@/app/view-data/components/common";
-import { evaluateLabInfoData } from "./labsService";
+import { evaluateLabInfoData, isLabReportElementDataList } from "./labsService";
 import { ConditionSummary } from "@/app/view-data/components/EcrSummary";
 import React from "react";
 import { toTitleCase } from "../utils/format-utils";
 import { formatDate, formatStartEndDateTime } from "./formatDateService";
 import { LabAccordion } from "../view-data/components/LabAccordion";
-import { isLabReportElementDataList } from "../utils/lab-utils";
 
 /**
  * Evaluates and retrieves patient details from the FHIR bundle using the provided path mappings.
