+++ containers/ecr-viewer/src/app/services/labsService.tsx
@@ -45,6 +45,25 @@ export interface LabReportElementData {
 }
 
 /**
+ * Checks if a given list is of type LabReportElementData[].
+ * Used to determine how to render lab results.
+ * @param labResults - Object to be checked.
+ * @returns True if the list is of type LabReportElementData[], false otherwise.
+ */
+export const isLabReportElementDataList = (
+  labResults: DisplayDataProps[] | LabReportElementData[],
+): labResults is LabReportElementData[] => {
+  const asLabReportElementList = labResults as LabReportElementData[];
+  return (
+    asLabReportElementList &&
+    asLabReportElementList.length > 0 &&
+    asLabReportElementList[0].diagnosticReportDataItems !== undefined &&
+    asLabReportElementList[0].organizationId !== undefined &&
+    asLabReportElementList[0].organizationDisplayDataProps !== undefined
+  );
+};
+
+/**
  * Extracts an array of `Observation` resources from a given FHIR bundle based on a list of observation references.
  * @param report - The lab report containing the results to be processed.
  * @param fhirBundle - The FHIR bundle containing related resources for the lab report.
